Separate aalwines lib from main exe
