+++ src/aalwines/model/Network.cpp
