+++ src/aalwines/model/NetworkWeight.h
