+++ src/aalwines/model/Query.cpp
