+++ src/aalwines/model/RoutingTable.cpp
@@ -26,7 +26,7 @@
 
 #include "RoutingTable.h"
 #include "Router.h"
-#include "utils/errors.h"
+#include "aalwines/utils/errors.h"
 #include "Network.h"
 
 #include <algorithm>
