+++ src/aalwines/model/builders/JuniperBuilder.h
@@ -27,8 +27,8 @@
 #ifndef JUNIPERBUILDER_H
 #define JUNIPERBUILDER_H
 
-#include "model/Router.h"
-#include "model/Network.h"
+#include "aalwines/model/Router.h"
+#include "aalwines/model/Network.h"
 
 #include <rapidxml.hpp>
