+++ src/aalwines/model/filter.cpp
