+++ src/aalwines/model/filter.h
