+++ src/aalwines/query/parsererrors.h
@@ -29,7 +29,7 @@
 #include <string>
 #include <stdexcept>
 #include <sstream>
-#include "utils/errors.h"
+#include "aalwines/utils/errors.h"
 
 namespace aalwines {
