+++ src/main.cpp
@@ -21,24 +21,20 @@
 // Created by Peter G. Jensen
 //
 
-#include "utils/errors.h"
-#include "query/QueryBuilder.h"
+#include <aalwines/utils/errors.h>
+#include <aalwines/query/QueryBuilder.h>
 
-#include "model/builders/JuniperBuilder.h"
-#include "model/builders/PRexBuilder.h"
+#include <aalwines/model/builders/JuniperBuilder.h>
+#include <aalwines/model/builders/PRexBuilder.h>
 
-#include "model/Router.h"
-#include "model/Network.h"
-#include "model/NetworkPDAFactory.h"
+#include <aalwines/model/NetworkPDAFactory.h>
 
-#include "query/parsererrors.h"
+#include <aalwines/query/parsererrors.h>
 #include <pdaaal/PDAFactory.h>
 #include <pdaaal/Solver_Adapter.h>
 #include <pdaaal/Reducer.h>
-#include "utils/stopwatch.h"
-#include "utils/outcome.h"
-
-#include <ptrie/ptrie_map.h>
+#include <aalwines/utils/stopwatch.h>
+#include <aalwines/utils/outcome.h>
 
 #include <boost/program_options.hpp>
 
@@ -46,8 +42,6 @@
 #include <string>
 #include <vector>
 #include <fstream>
-#include <memory.h>
-#include <boost/filesystem/path.hpp>
 
 namespace po = boost::program_options;
 using namespace aalwines;
